refactor(ddd): Modelsディレクトリ配下のクラスを Domain, Infrastructure, ViewModels に再配置
