+++ Controllers/FileController.cs
@@ -7,7 +7,7 @@ namespace SuumoScraping.Controllers
     using System.Threading.Tasks;
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.Extensions.Logging;
-    using SuumoScraping.Models;
+    using SuumoScraping.Domain.Models;
     using SuumoScraping.UseCases;
     using SuumoScraping.ViewModels;
