+++ Controllers/HomeController.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using SuumoScraping.Models;
+using SuumoScraping.ViewModels;
 
 namespace SuumoScraping.Controllers
 {
