+++ Views/_ViewImports.cshtml
@@ -1,5 +1,6 @@
-﻿@using SuumoScraping
-@using SuumoScraping.Models
+@using SuumoScraping
+@using SuumoScraping.Domain.Models
+@using SuumoScraping.ViewModels
 @using SuumoScraping.Extensions
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
